+++ src/components/DisplayResults.js
@@ -7,12 +7,6 @@ class DisplayResults extends Component {
         super();
         this.state = {
             allEvents: [],
-            filteredEvents: [],
-            isLoading: true,
-            eventCategories: [],
-            userCategory: "",
-            eventVenues: [],
-            userVenue: "",
             user1: {},
             user2: {},
             currentUser: "1",
@@ -20,83 +14,6 @@ class DisplayResults extends Component {
             // App.js passes info to DisplayResults
             //match value to user1 or 
         }
-    }
-    // converting time function to string so it can be passed as a number in template literals 
-    formatDate = (dateObject, time) => {
-        let year = dateObject.getFullYear();
-        let month = dateObject.getMonth() + 1;
-        if (month < 10) {
-            month = '0' + month;
-        }
-        let day = dateObject.getDate();
-        if (day < 10) {
-            day = '0' + day;
-        }
-        let hours = time.substr(0, 2);
-        console.log(hours);
-        let minutes = time.substr(3, 5);
-        console.log(minutes);
-        let dateString = `${year}-${month}-${day}T${hours}:${minutes}:00`;
-        return dateString
-    }
-
-    //time input format localStartEndDateTime=2019-06-05T17:00:00,2019-06-05T20:00:00 
-    getTicketmasterData = (location, startDate, endDate) => {
-        console.log('parameters that go to the API Call')
-        console.log("location: ", location)
-        console.log("Start Date: ", startDate);
-        console.log("End Date: ", endDate);
-        axios({
-            method: "GET",
-            url: `https://app.ticketmaster.com/discovery/v2/events.json?apikey=cpqJuV2A3YqkXOJylkTrDzVGLRKZ5hp5&city=${location}&localStartEndDateTime=${startDate},${endDate}`,
-            dataResponse: "jsonp",
-        }).then((response) => {
-            // console.log(response);
-            response = response.data._embedded.events;
-            console.log(response)
-            // console.log(allEvents)
-            this.setState({
-                // allEvents is the good return we never modify
-                allEvents: response,
-                // filteredEvents is it item we want to modify based on user input
-                filteredEvents: response,
-                isLoading: false,
-                
-            })
-            this.getEventCategories();
-            this.getEventVenues();
-        })
-    }
-    getEventCategories = () => {
-        // loop through all events 
-        let eventCategories = ["All Categories"];
-        for (let i = 0; i < this.state.allEvents.length; i ++) {
-            let eventCategory = this.state.allEvents[i].classifications[0].segment.name;
-            console.log(eventCategory);
-            if (!eventCategories.includes(eventCategory)) {
-                eventCategories.push(eventCategory)
-            }
-        }
-        console.log(eventCategories);
-        this.setState({
-            // ... copies the items to the array
-            eventCategories: [...eventCategories],
-        })
-    }
-    getEventVenues = () => {
-        // loop through all events 
-        let eventVenues = ["All Venues"];
-        for (let i = 0; i < this.state.allEvents.length; i++) {
-            let eventVenue = this.state.allEvents[i]._embedded.venues[0].name;
-            if (!eventVenues.includes(eventVenue)) {
-                eventVenues.push(eventVenue)
-            }
-        }
-        console.log(eventVenues);
-        this.setState({
-            // ... copies the items to the array
-            eventVenues: [...eventVenues],
-        })
     }
 
     filterEvents = (e, textFilter, categoryDropdown, venueDropdown) => {
@@ -114,7 +31,7 @@ class DisplayResults extends Component {
                 if (categoryDropdown==="All Categories"){
                     return true;
                 }
-                else{
+                else {
                     return eventObj.classifications[0].segment.name.includes(categoryDropdown);
                 }
             })
@@ -130,30 +47,17 @@ class DisplayResults extends Component {
             filteredEvents: [...filteredEvents],
         })
     }
-    // default is user one
-    // which is the selected
-    // current user will = user1
+
+
     handleChangeRadio = (event) => {
             this.setState({
+                //sets currentUser to be either User1 or User2 on toggle
                 currentUser: event.target.value
             })
         }
-    componentDidMount() {
-        let date = this.props.date;
-        let timeStart = this.props.timeStart;
-        let timeEnd = this.props.timeEnd;
-        let startDateTime = this.formatDate(date, timeStart);
-        let endDateTime = this.formatDate(date, timeEnd);
-        let location = this.props.location;
-        this.getTicketmasterData(location, startDateTime, endDateTime);  
-    }
-
-
 
     render() {
         return (
-            this.state.isLoading ? <h1>Getting Your Events...</h1> :
-
                 <div className="display-events">
                     <div className="display-content">
                         {/* errorhandling here : write a condition if city name is found in API, do below. else "Please enter a valid city name and date/time range*/}
